% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_functions.R
\name{boundaries}
\alias{boundaries}
\alias{boundaries_all}
\alias{boundaries_ols}
\alias{boundaries_ols_XonY}
\alias{boundaries_RMA}
\alias{boundaries_MM}
\alias{boundaries_AdjQ}
\alias{boundaries_medians}
\alias{boundaries_q75}
\title{Functions to calculate boundary nutrient concentrations}
\usage{
boundaries_all(data, yvar = "y.u", xvar = "x.u", status_boundaries = c(HG
  = 0.8, GM = 0.6, MP = 0.4), class_var = "status")

boundaries_ols(data, yvar = "y.u", xvar = "x.u", status_boundaries = c(HG
  = 0.8, GM = 0.6, MP = 0.4))

boundaries_ols_XonY(data, yvar = "y.u", xvar = "x.u",
  status_boundaries = c(HG = 0.8, GM = 0.6, MP = 0.4))

boundaries_RMA(data, yvar = "y.u", xvar = "x.u", status_boundaries = c(HG
  = 0.8, GM = 0.6, MP = 0.4))

boundaries_MM(data, yvar = "y.u", xvar = "x.u", status_boundaries = c(HG =
  0.8, GM = 0.6, MP = 0.4))

boundaries_AdjQ(data, yvar = "y.u", xvar = "x.u", class_var = "status")

boundaries_medians(data, yvar = "y.u", xvar = "x.u", class_var = "status")

boundaries_q75(data, yvar = "y.u", xvar = "x.u", class_var = "status")
}
\arguments{
\item{data}{Appropriate water quality data for estimating nutrient status boundaries}

\item{yvar}{The name of the response variable, typically EQR. Defaults to y.u}

\item{xvar}{The name of the pressure variable, e.g. total phosphorus. Defaults to x.u}

\item{status_boundaries}{A named vector of status boundaries on the scale of \code{yvar}. 
Default contains 0.8, 0.6 and 0.4 for the High-Good, Good-Moderate and Moderate-Poor boundaries}

\item{class_var}{name of variable defining status classes. Defaults to "status".}
}
\value{
a named vector of estimated nutrient boundaries on the scale of \code{xvar}
}
\description{
Functions to calculate boundary nutrient concentrations
}
\section{Functions}{
\itemize{
\item \code{boundaries_all}: A wrapper function to estimate boundaries using all methods

\item \code{boundaries_ols}: Estimates nutrient boundaries using ordinary-least-squares

\item \code{boundaries_ols_XonY}: Estimates nutrient boundaries using ordinary-least-squares

\item \code{boundaries_RMA}: Estimates nutrient boundaries using Ranged Major Axis regression.

\item \code{boundaries_MM}: Estimates nutrient boundaries using the "mismatch" method of minimising the difference between two types of missclassification.

\item \code{boundaries_AdjQ}: Estimates nutrient boundaries using the "adjusted quartiles" method.

\item \code{boundaries_medians}: Estimates nutrient boundaries as the mean of the median concentrations in adjacent classes

\item \code{boundaries_q75}: Estimates nutrient boundaries as the upper quartile of the relevant status class
}}

